MAO/Version-2.2/Code/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.0
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
autopep8==1.6.0
bidict==0.22.1
blinker==1.6.2
certifi==2023.5.7
cffi==1.16.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.1
cryptography==41.0.5
cycler==0.12.1
decorator==4.4.2
dill==0.3.6
distlib==0.3.6
docstring-to-markdown==0.11
easy-dict==0.1.1
easydict==1.10
et-xmlfile==1.1.0
faiss-cpu==1.7.4
filelock==3.12.2
flake8==6.0.0
Flask==2.3.2
Flask-SocketIO==5.3.4
fonttools==4.43.1
frozenlist==1.3.3
fsspec==2023.10.0
gitdb==4.0.10
GitPython==3.1.37
graphviz==0.20.1
greenlet==3.0.1
h11>=0.13,<0.15
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1

huggingface-hub==0.19.4
hupper==1.12
hyperframe==5.2.0
identify==2.5.24
idna==2.10
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
jsonpickle==3.0.2
kiwisolver==1.4.5
lazy-object-proxy==1.9.0
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.8.0
mccabe==0.7.0
mistune==3.0.1
more-itertools==10.1.0
moviepy==1.0.3
msgpack==1.0.7
multidict==6.0.4

nbconvert==7.8.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numba==0.58.1
numpy==1.24.3
openai==1.63.0
openpyxl==3.1.2
packaging==23.0
pandas==2.1.1
parso==0.8.3
PasteDeploy==3.0.1
patsy==0.5.3
Pillow==9.5.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.5.3
player==0.6.1
plotly==5.17.0
pluggy==1.2.0
pooch==1.8.0
pre-commit==3.3.3
proglog==0.1.10
pycodestyle==2.10.0
pycparser==2.21
pydocstyle==6.2.3
pyflakes==3.0.1
pygame==2.5.1
pylint==2.17.0
pyparsing==3.1.1
pyramid==2.0.2
pyramid-jinja2==2.10
PySocks==1.7.1
pysrt==1.1.2
python-dateutil==2.8.2
python-engineio==4.5.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.7.1
python-socketio==5.8.0
pytoolconfig==1.2.5
pytz==2023.3.post1
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
retrying==1.3.4
rfc3986==1.5.0
rope==1.7.0
safetensors==0.4.0
schedule==1.2.1
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0

smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.23
statsmodels==0.14.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
timeout-decorator==0.5.0
tokenizers==0.15.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.65.0
transformers==4.35.2
translationstring==1.4
httpx==0.28.1
httpcore==1.0.4
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.3
venusian==3.0.0
virtualenv==20.23.0
webcolors==1.13
WebOb==1.8.7
Werkzeug==2.3.6
whatthepatch==1.0.4
wrapt==1.15.0
yapf==0.32.0
yarl==1.9.2
zipp==3.16.2
zope.deprecation==5.0
zope.interface==6.0
pydotplus==2.0.2
python_Levenshtein==0.27.1
six==1.16.0
lxml
requests
flask
openpyxl
pyyaml
tqdm
scikit-learn
scipy
altair
pm4py